requirements.txt
pywin32==302
pywinauto==0.6.8
PyAutoGUI==0.9.53
urllib3==1.26.7
psutil==5.8.0
keyboard==0.13.5
mouse==0.7.1

selenium==4.1.0
Eel==0.12.4
Flask~=2.0.3
websocket-server==0.6.4
requests==2.26.0
aiohttp~=3.8.1
python-dotenv==0.19.2
winshell==0.6
schedule~=1.1.0
openpyxl~=3.0.10
xlrd==2.0.1